#3 & #4 create route subfolders: one route file for each Object/Table
